WIP attempt to add firebase unit tests.
Committing to remember what I tried, but have decided it's not worth the
effort right now. Having a lot of trouble getting the firebase emulator
working locally - got to a point where I needed to set up the proper
version of JDK just to get the emulator running.

The emulator is still in Beta mode, so I don't think it's worth spending
more time on at this point:
https://firebase.google.com/docs/rules/emulator-setup