Make firebase cli SDK play nice with gatsby.
This article describes the problem
https://kyleshevlin.com/firebase-and-gatsby-together-at-last

TLDR Issues during `gatsby build` because gatsby does server side rendering,
and the client SDK has module dependencies that only work in non-server environement.

This plugin does the same thing as the solution described there.

Might be worth wrapping things in a context that provides `firebase`
everywhere, like in this example:
https://github.com/muhajirdev/gatsby-starter-firebase