WIP Experiment with manipulating the d3 hierarchy.
Goal is to not render the leaf nodes until you've zoomed in a few
levels to keep the total number of DOM elements managable.

Current problem: When removing the leaf nodes, it's hard to put them
back. It's hard to maintain a the original hierachy, because its so
nested and self-referential that even mutating the copy acts on the
original. 🤔