Fix node picking cross-browser bug.
The UIEvent.layerX and layerY properties are not standardized
across browsers. This was made apparent as the mouse click coordinates
were wildly different between chrome and firefox, and was breaking node
picking and zooming on firefox.

https://developer.mozilla.org/en-US/docs/Web/API/UIEvent/layerX

Solution was to switch to d3's mouse click coordinate accessor, which
is consistent between browsers.

Also:
- Some updates to when labels are rendered. Generally, now only render
  the layer 1 down from focus.