[WIP] Visualization Experiments (#7)
* Try out d3 packed circles visualization.

* Add bubble chart experiment.

Some thoughts:
- The packed circles hierarchy wastes too much space when breaking it
down into so many levels. Probably will need to drop the postal/zip code
level - breakdown by country should be enough locality.
- Both the packed circles and flat bubble charts will become too
detailed (too many SVG circles and labels) and will be sluggish once
we get up to thousands of commitments. We'll need to pre-process the data
somehow. Perhaps aggregating 10s or hundreds of commitments into one
bubble with the value 10 or 100.

* Experiment with dynamically updating the circle pack data hierarchy.

Proves we can modify the data hierarchy on the fly.
This would allow us do thing like:
- Only render the individual commitments when the parent is zoomed in
  on.

Important:
- Must use a data join key so the DOM elements can consistently map to
  the data hierarchy when things change dynamically.
- Figured out how do transitions

* WIP Experiment with manipulating the d3 hierarchy.

Goal is to not render the leaf nodes until you've zoomed in a few
levels to keep the total number of DOM elements managable.

Current problem: When removing the leaf nodes, it's hard to put them
back. It's hard to maintain a the original hierachy, because its so
nested and self-referential that even mutating the copy acts on the
original. 🤔

* Experiment with Canvas bubble chart => faster.

The canvas bubble chart renders thousands of circles much quicker than
the SVG. Next steps are to try adding in basic animation, which will
enable things like zooming.

* Add zooming animation with canvas.

Relatively smooth up to 5K circles.

* Implement node picking and zooming.

Based on:
- http://bl.ocks.org/nbremer/db24422abdb20150a9dd
- https://bocoup.com/blog/2d-picking-in-canvas

* Get canvas hierarchical packed circles working.

- Make the same canvas render logic work for flat and hierarchical data.
- Move dummy data generators to a helper file.

* Clean up: Remove SVG versions, show both flat and hierarchical canvas.

Also:
- Fix zooming on root node
- Fix problems with switching between flat and hierarchical versions
  with useEffect cleanups: stop timer, remove callbacks.

* Add name labels to commitment visualization.

Also:
- Some perf tweaks: Only render labels after the animation is done and
  when focused on the leaf nodes.
- Random name generation in the dummy data.

* Dynamically resize.

- Create a wrapper component that manages detecting and maintaining
  dimensions via a ref in the child node (div wrapping the viz).
- Learned the ref must be embedded in a child that is a HTML element
  component (i.e. had to put the ref in a wrapper div, not in the
  Viz component)
- Fix multiple rendering bug due to repeated append operations in the
  render function.

* Cleanup.

* Cleanup + move mock data generation out of viz component.

* WIP. First pass at dynamically re-arranging data hierarchy.

* Fix node picking cross-browser bug.

The UIEvent.layerX and layerY properties are not standardized
across browsers. This was made apparent as the mouse click coordinates
were wildly different between chrome and firefox, and was breaking node
picking and zooming on firefox.

https://developer.mozilla.org/en-US/docs/Web/API/UIEvent/layerX

Solution was to switch to d3's mouse click coordinate accessor, which
is consistent between browsers.

Also:
- Some updates to when labels are rendered. Generally, now only render
  the layer 1 down from focus.

* Add rudimentary hierarchy key UI controls.

* Forgot to commit test data file.

* Basic styling for hierarchy controls.

* Add note about performance problem.

* Render optimizations + Comments.

Noticed the animation was continuously running and using a lot of cycles
needlessly. It should only run when there is a change (zoom + refocus).

This change starts the animation on click to zoom.

Committing now with a lot TODOs and thoughts as comments before cleaning
up.

* Remove old dummy data generators.

* Write up thoughts on picking bug: Anti-aliasing.

* Clean out console logs.

* Fix (compromise) on anti-aliasing color-pick problem.

Problem: If we want to pick a deep node (e.g. leaf) from a high level (e.g. root)
focus, the small leaf nodes are anti-aliased so the pixel chosen is often a
blended color rather than the pickColor. This results in zooming to the
wrong node (if the anti-aliased pixel color matches an existing node) or
no node getting picked.

Attempts that didn't work:
- Turn off anti-aliasing when drawing the context. Seemed to help a bit,
  but not enough.
- Sort the nodes in the d3 hierarchy using sortkeys composed of all the
  hierarachy keys concatenated. So the pick colors of nodes in proximity
  would be more similar to eachother than unrelated/far-away nodes. This
  also didn't seem to help (though I'm honestly not sure I actually got
  it working properly before deciding to switch to my final, easier,
  idea)

Compromise Solution: Only allow zooming 1 level down at a time.
- Cons: You cant skip multiple levels and get right to the node you want
  to click.
- Pros: Avoids the anti-aliasing problem of small circles, because you
  never make such a large jump. And I think thats the happy path anyway,
  no one actually cares about a specific low level node (one can't tell
  what they are from a high level anyway), so this forces a sensible
  and informative exploration.

* Delete flat vis + move HierarchyControls to its own module.

- Flat vis is irrelevant now that we have UI control over the data
  hierarchy

* Rename: BubbleChart -> PackedCircles

* Refactor: Renaming, comments, pull logic into helpers.

* Remove unneeded global dt + variable renames and organizing.

* Make interpolation function more functional.

Starting on the process of making these functions more functional.
There are so many global (in the file) variables getting read from and
written to from so many different places it makes it very hard to figure
out whats going on.

* Remove animation time tracking global.

* Move render state out of global vars.

* Fix hierarchy controls bug.

* Remove deprecated flat viz link.