Commitment flow form: Real Submission + Validation (#13)
* Move StepNavigation control inside flow components.

Why: To allow inserting extra actions to fire on 'Next' for specific
steps. Specifically, to allow form submission to happen when the user
is on the Signup component and hits Next.

Also:
- Add additional fields to the form
- Misc formatting changes made by Prettier

* Diagnosed loss-of focus on form input issue.

Temp fix here with comment explaining the issue and next steps.

* Implement basic form input management with Formik.

* Connect commitment submission form to backend!

* WIP Getting stuck, need to switch tacts.

Getting stuck trying to inject the Formik validate->submit handler
flow into a beforeNext function to run inside the step handler. Its
getting quite complicated and async.

Probably better to switch tactics. Do we really need to have the same
StepNavigator component fit the needs of every part of the flow? The
SignUp component is fundamentally different in that Next is really
Submit, so having a different interface would probalby make more sense.

* Fix a bunch of Frontend + Backend validation issues.

- Update step navigation to make it possible to submit on the
  Next button for Signup page.
- Add JSON validation schema restrictions (for the backend)
- Fix the divestment and participate commitment state variable names
- Other misc stuff

* Make ReviewSubmission more readable.

Obviously it needs real design, but this is a better temp solution.

* Deprecate experimental firebase submission page.

Now we have the real thing in the commitments flow.